remove wasi-vfs dependency
It turns out that wasi-vfs isn't needed after all -- our Wizer initialization
step is sufficient to allow CPython to load everything it needs from the host
filesystem.  As a bonus, skipping the wasi-vfs step ensures that the module size
is automatically minimized -- no need to use `rsync` for staging and excluding
files by hand.

One possible drawback is that an application might try to dynamically load
modules from the filesystem at runtime (either built-in Python modules or
otherwise).  In that case, we may need to either use wasi-vfs or give the
developer some other way to pull the needed modules in at Wizer time (e.g. via
an app-provided init function which is called from the Wizer init function).

Signed-off-by: Joel Dice <joel.dice@fermyon.com>