rearrange repo and add `spin-python-cli`
`spin-python-cli` is a stand-alone binary for converting a Python app to a
Spin-compatible Wasm module.  The intention is that we'll publish this as a
`Spin` plugin, analogous to the JS/TS SDK's `js2wasm` utility.

The implementation is a bit more complicated than `js2wasm` because the Python
interpreter loads the core library, the app, and its dependencies from disk, so
we have to:

- Embed the core library in the binary and extract it to a temporary directory prior to pre-initializing with Wizer
- Search for the `pipenv` `site-packages` directory and add it to `PYTHONPATH`
- Set various environment variables so the Python interpreter knows where to find things (and not buffer stdio)
- Set another environment variable so the pre-init function knows which module the app lives in

Signed-off-by: Joel Dice <joel.dice@fermyon.com>