call `__wasm_call_ctors` in all exported functions
Now that we've updated to the latest Rust Spin SDK, which uses a newer
`wit-bindgen`, we need to call `__wasm_call_ctors` explicitly in order to
initialize `wasi-libc` (e.g. grab environment variables, etc.).  Previously,
this was done implicitly, but that caused other problems, so `wit-bindgen` does
it explicitly, which has the side effect of telling `wasm-ld` _not_ to do it
implicitly.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>