add docs
This adds a set of stub modules which mirror the API generated from
`spin-python-engine`'s Rust code by PyO3, from which we can generate HTML
documentation to host on GitHub Pages.  Ideally, we'd generate those docs from
the Rust code using some combination of `PyO3` and `maturin`, but I'm not sure
how to do that, and we're planning to deprecate this whole project in favor of
`componentize-py` and a new SDK based on it, so I went the easy route.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>