fix outbound request body streaming (#103)
There was a subtle issue in `spin_sdk.http.send_async` which caused deadlock
when sending a request to a server that waits for the full request body before
sending any part of a response.  Specifically, we must close the `Sink` as soon
as we've written the whole body in order to communicate that there will be no
further bytes.  Previously, we only closed the `Sink` after we'd started
receiving the response, which can lead to deadlock.

This also bumps the package version to 3.2.0 and makes a minor README.md
correction.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>