make `spin_sdk.http.IncomingHandler` optional
When building a component intended for composition with other components, it may
make sense to target the `spin-imports` world defined in the
`src/spin_sdk/wit/spin.wit` file in this repo.  However, that currently won't
work because `spin_sdk.http` unconditionally imports `spin_sdk.wit.exports`,
which won't exist when targeting `spin-imports`.

The solution is to catch the `ImportError` if importing the `exports` submodule
(or its `IncomingHandler` class) fails and just skip defining
`spin_sdk.http.IncomingHandler` when that happens.

Fixes #110

Signed-off-by: Joel Dice <joel.dice@fermyon.com>